complete 4.21*

+++ part4/bloglist-backend/controllers/blogs.js
@@ -49,11 +49,23 @@ blogRouter.post('/', async (req, res, next) => {
 
 // route to delete a post
 blogRouter.delete('/:id', async (req, res, next) => {
+    const token = getTokenFrom(req)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
     try {
-        await Blog.findByIdAndRemove(req.params.id)
-        res.status(204).end()
-    } catch (exception) {
-        next(exception)
+        const blog = await Blog.findById(req.params.id)
+        console.log('her ei s blog', blog)
+        if (blog.user.toString() === decodedToken.id) {
+            await Blog.findByIdAndRemove(req.params.id)
+            res.status(204).end()
+        } else {
+            res.send('Not authorized to delete')
+        }
+    } catch (error) {
+        next(error)
     }
 })
